+++ packages/fabric-core/src/plugins/index.ts
@@ -2,14 +2,4 @@ export { barrelPlugin } from './barrelPlugin.ts'
 export { createPlugin } from './createPlugin.ts'
 export { fsPlugin } from './fsPlugin.ts'
 export { graphPlugin } from './graphPlugin.ts'
-export { progressPlugin } from './progressPlugin.ts'
 export { loggerPlugin } from './loggerPlugin.ts'
-export type {
-  LoggerPluginOptions,
-  LoggerPluginState,
-  LoggerEventMessage,
-  LoggerStatusMessage,
-  LoggerWelcomeMessage,
-  LoggerHistoryMessage,
-  LoggerMessage,
-} from './loggerPlugin.ts'